Flow control spec cleanup: control frames.
Delete the section of the spec that mandates behavior on control frame
sizes. It is wrong to insist on sending a RST_STREAM and not processing
the control frame, since control frames may lead to session errors. For
example, if the header block is malformed, the compression context is
busted, which is a session error, and the receiver must detect that.
It's not safe to proceed by ignoring the control frame and sending a
RST_STREAM.

Implementations can either choose to introduce head of line blocking
until memory becomes available to process the control frame in its
entirety, or they can close the connection with a session error.
Removing from the spec and leaving this behavior up to implementation.